third_party/WebKit/LayoutTests/fast/dom/Range/range-intersectsNode-exception-expected.txt

Test for Bug: 75799 - Calling intersectsNode on a detached range should throw

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS range.intersectsNode(document.getElementById('b1')) threw exception InvalidStateError: Failed to execute 'intersectsNode' on 'Range': The range has no container. Perhaps 'detatch()' has been invoked on this object?.
PASS successfullyParsed is true

TEST COMPLETE
